Add template export functionality to the new login component. Export existing templates without having to log in.
